chore: Move crate::types::Variable to dtl_lexer::types::Variable
