+++ dtl-lexer/src/types.rs
@@ -1,3 +1,5 @@
+use crate::TemplateContent;
+
 #[derive(Clone, Copy)]
 pub struct TemplateString<'t>(pub &'t str);
 
@@ -20,5 +22,61 @@ pub trait IntoTemplateString<'t> {
 impl<'t> IntoTemplateString<'t> for &'t str {
     fn into_template_string(self) -> TemplateString<'t> {
         TemplateString(self)
+    }
+}
+
+pub struct PartsIterator<'t> {
+    variable: &'t str,
+    start: usize,
+}
+
+impl<'t> Iterator for PartsIterator<'t> {
+    type Item = (&'t str, (usize, usize));
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.variable.is_empty() {
+            return None;
+        }
+
+        match self.variable.find('.') {
+            Some(index) => {
+                let part = &self.variable[..index];
+                let at = (self.start, index);
+                self.start += index + 1;
+                self.variable = &self.variable[index + 1..];
+                Some((part, at))
+            }
+            None => {
+                let part = self.variable;
+                self.variable = "";
+                Some((part, (self.start, part.len())))
+            }
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct Variable {
+    pub at: (usize, usize),
+}
+
+impl<'t> Variable {
+    pub fn new(at: (usize, usize)) -> Self {
+        Self { at }
+    }
+
+    pub fn parts(
+        &self,
+        template: TemplateString<'t>,
+    ) -> impl Iterator<Item = (&'t str, (usize, usize))> {
+        let start = self.at.0;
+        let variable = template.content(self.at);
+        PartsIterator { variable, start }
+    }
+}
+
+impl<'t> TemplateContent<'t> for Variable {
+    fn content(&self, template: TemplateString<'t>) -> &'t str {
+        template.content(self.at)
     }
 }
