+++ src/types.rs
@@ -1,36 +1,6 @@
 use num_bigint::BigInt;
 
-use dtl_lexer::types::TemplateString;
-
-struct PartsIterator<'t> {
-    variable: &'t str,
-    start: usize,
-}
-
-impl<'t> Iterator for PartsIterator<'t> {
-    type Item = (&'t str, (usize, usize));
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.variable.is_empty() {
-            return None;
-        }
-
-        match self.variable.find('.') {
-            Some(index) => {
-                let part = &self.variable[..index];
-                let at = (self.start, index);
-                self.start += index + 1;
-                self.variable = &self.variable[index + 1..];
-                Some((part, at))
-            }
-            None => {
-                let part = self.variable;
-                self.variable = "";
-                Some((part, (self.start, part.len())))
-            }
-        }
-    }
-}
+use dtl_lexer::types::Variable;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct Text {
@@ -51,26 +21,6 @@ pub struct TranslatedText {
 impl TranslatedText {
     pub fn new(at: (usize, usize)) -> Self {
         Self { at }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct Variable {
-    pub at: (usize, usize),
-}
-
-impl<'t> Variable {
-    pub fn new(at: (usize, usize)) -> Self {
-        Self { at }
-    }
-
-    pub fn parts(
-        &self,
-        template: TemplateString<'t>,
-    ) -> impl Iterator<Item = (&'t str, (usize, usize))> {
-        let start = self.at.0;
-        let variable = template.content(self.at);
-        PartsIterator { variable, start }
     }
 }
